Fix warnings about casting pointer from/to different size integer
These appear under 64-bit Windows. Only the sciwrappers.c warning is
potentially dangerous. For win32.c, the "handle" provides some useful
information, while "lStdHandle" does not.

+++ tagmanager/src/tm_workspace.c
@@ -398,7 +398,7 @@ static guint tm_file_inode_hash(gconstpointer key)
 #ifdef TM_DEBUG
 		g_message ("Hash for '%s' is '%d'\n", filename, file_stat.st_ino);
 #endif
-		return g_direct_hash ((gpointer)(gulong)file_stat.st_ino);
+		return g_direct_hash ((gpointer)(intptr_t)file_stat.st_ino);
 	} else {
 		return 0;
 	}
